fix: rename UpdateBench* -> UpdateReleaseGroup* - fix import routes for previous commit
